+++ examples/tasks.py
@@ -4,16 +4,11 @@ import os
 from loguru import logger
 
 from rearq import ReArq, Task
-from rearq.server import models
 
-config = {
-    "connections": {
-        "default": f"mysql://root:{os.getenv('MYSQL_PASS') or '123456'}@127.0.0.1:3306/rearq"
-    },
-    "apps": {"models": {"models": [models], "default_connection": "default"}},
-    "use_tz": True,
-}
-rearq = ReArq(tortoise_config=config, delay_queue_num=2)
+rearq = ReArq(
+    db_url=f"mysql://root:{os.getenv('MYSQL_PASS') or '123456'}@127.0.0.1:3306/rearq",
+    delay_queue_num=2,
+)
 
 
 @rearq.on_shutdown