+++ rearq/server/routes/worker.py
@@ -26,7 +26,10 @@ async def get_workers(request: Request, rearq: ReArq = Depends(get_rearq)):
             await JobResult.filter(worker=worker_name)
             .annotate(count=Count("job_id"))
             .group_by("job__status")
-            .values("count", status="job__status",)
+            .values(
+                "count",
+                status="job__status",
+            )
         )
         item = {"name": worker_name, "job_stat": {job["status"]: job["count"] for job in job_stat}}
         item.update(json.loads(value))